feat(desktop): resume Claude sources from a stored snapshot
Wire the worker's evidence pass to attempt visit_claimed_resumed for
each Claude file source instead of always reading from the start.

Add VendorAdapter::empty_resume_state, letting an adapter-agnostic
caller build a bootstrap StreamSnapshot (offset zero, fresh adapter
state) for a source with no stored resume yet. Passing this to
visit_claimed_resumed reads the whole source, matching visit_claimed,
while still producing a genuine AdapterResume for the next pass -
this is how a source becomes resumable from its first pass. Only
ClaudeAdapter overrides it; other adapters keep reading in full.

Add TurnRowStore::delete_rows_for_source, used when a resumed visit
reports SourceChanged after already flushing some new rows: the
retried full read starts a fresh row sequence and must not join or
duplicate the partial one, and the turn table has no unique
constraint that would prevent that on its own.

Thread SourcePublishMode and the resulting StoredResume out of
stream_vendor_with_hooks as EvidencePass::source_outcomes, through to
Store::publish_projections via insights_worker::apply_outcome,
replacing the placeholder empty slice from the prior commit.

Split analysis.rs and insights_worker.rs's inline test modules into
sibling analysis/tests.rs and insights_worker/tests.rs files (no
content changes beyond the move) to stay under the file-size cap
after these additions.
