+++ apps/desktop/src-tauri/src/store/tests/resume_tests.rs
@@ -149,8 +149,9 @@ fn a_resumed_source_re_stamps_only_its_own_appended_rows() {
     let first_published_fence = store.evidence(&key).unwrap().unwrap().published_fence;
 
     // Second pass: the session resumes. Only the parent appends one new row
-    // under the new claim fence; the child is untouched — no new rows, no
-    // entry in `sources`.
+    // under the new claim fence; the child writes nothing new but is still
+    // named as Resumed — a still-present, unchanged source a resume-aware
+    // adapter names on every visit regardless of new row count.
     mark_evidence_pending_in(&store.lock(), &key).unwrap();
     let next_claim = store
         .claim_next_evidence(&["claude-code"], 200, 60)
@@ -166,11 +167,18 @@ fn a_resumed_source_re_stamps_only_its_own_appended_rows() {
         .write_turn_rows(&[turn_row_for("resume-restamp", 1)])
         .unwrap();
     let next_completion = evidence_completion(&next_claim, PublishedEvidence::Ready, "{}".into());
-    let sources = [SourcePublishOutcome {
-        source_key: "resume-restamp".into(),
-        mode: SourcePublishMode::Resumed,
-        resume: None,
-    }];
+    let sources = [
+        SourcePublishOutcome {
+            source_key: "resume-restamp".into(),
+            mode: SourcePublishMode::Resumed,
+            resume: None,
+        },
+        SourcePublishOutcome {
+            source_key: "child-1".into(),
+            mode: SourcePublishMode::Resumed,
+            resume: None,
+        },
+    ];
 
     assert!(
         store
@@ -225,7 +233,10 @@ fn a_full_read_source_replaces_only_its_own_published_rows() {
     );
 
     // Second pass: the parent forces a full read (a tail rewrite, say) and
-    // rewrites its rows from scratch; the child is untouched.
+    // rewrites its rows from scratch; the child writes nothing new but is
+    // still named as Resumed — a still-present, unchanged source a
+    // resume-aware adapter names on every visit regardless of new row
+    // count.
     mark_evidence_pending_in(&store.lock(), &key).unwrap();
     let next_claim = store
         .claim_next_evidence(&["claude-code"], 200, 60)
@@ -241,11 +252,18 @@ fn a_full_read_source_replaces_only_its_own_published_rows() {
         .write_turn_rows(&[turn_row_for("resume-full-replace", 0)])
         .unwrap();
     let next_completion = evidence_completion(&next_claim, PublishedEvidence::Ready, "{}".into());
-    let sources = [SourcePublishOutcome {
-        source_key: "resume-full-replace".into(),
-        mode: SourcePublishMode::Full,
-        resume: None,
-    }];
+    let sources = [
+        SourcePublishOutcome {
+            source_key: "resume-full-replace".into(),
+            mode: SourcePublishMode::Full,
+            resume: None,
+        },
+        SourcePublishOutcome {
+            source_key: "child-1".into(),
+            mode: SourcePublishMode::Resumed,
+            resume: None,
+        },
+    ];
 
     assert!(
         store
@@ -271,6 +289,90 @@ fn a_full_read_source_replaces_only_its_own_published_rows() {
         child_rows.len(),
         1,
         "the untouched child's row must survive, not be deleted"
+    );
+}
+
+#[test]
+fn a_vanished_source_has_its_rows_and_resume_dropped_on_the_next_publish() {
+    let store = store();
+    let (record, claim) = claimed_projection(&store, "resume-vanish", 100, 60);
+    let key = record.key.clone();
+    let writer = FencedTurnRowStore::new(store.clone(), key.clone(), claim.claim_fence);
+    writer
+        .write_turn_rows(&[turn_row_for("resume-vanish", 0), turn_row_for("child-1", 0)])
+        .unwrap();
+    let completion = evidence_completion(&claim, PublishedEvidence::Ready, "{}".into());
+    let sources = [
+        SourcePublishOutcome {
+            source_key: "resume-vanish".into(),
+            mode: SourcePublishMode::Full,
+            resume: Some(sample_resume("fp-parent")),
+        },
+        SourcePublishOutcome {
+            source_key: "child-1".into(),
+            mode: SourcePublishMode::Full,
+            resume: Some(sample_resume("fp-child")),
+        },
+    ];
+    assert!(
+        store
+            .publish_projections(&record, None, &completion, &[], &sources)
+            .unwrap()
+    );
+
+    // Second pass: the child transcript is gone (removed from disk, or
+    // unreadable this time), so this pass names and reads only the parent.
+    mark_evidence_pending_in(&store.lock(), &key).unwrap();
+    let next_claim = store
+        .claim_next_evidence(&["claude-code"], 200, 60)
+        .unwrap()
+        .expect("reclaimable");
+    let next_record = projection_record(
+        key.clone(),
+        "sv1:resume-vanish",
+        next_claim.source_generation,
+    );
+    let next_writer = FencedTurnRowStore::new(store.clone(), key.clone(), next_claim.claim_fence);
+    next_writer
+        .write_turn_rows(&[turn_row_for("resume-vanish", 1)])
+        .unwrap();
+    let next_completion = evidence_completion(&next_claim, PublishedEvidence::Ready, "{}".into());
+    let next_sources = [SourcePublishOutcome {
+        source_key: "resume-vanish".into(),
+        mode: SourcePublishMode::Resumed,
+        resume: None,
+    }];
+
+    assert!(
+        store
+            .publish_projections(&next_record, None, &next_completion, &[], &next_sources)
+            .unwrap()
+    );
+
+    let published = store.published_turn_rows(&key).unwrap().expect("ready");
+    let parent_rows: Vec<_> = published
+        .iter()
+        .filter(|row| row.source_key == "resume-vanish")
+        .collect();
+    let child_rows: Vec<_> = published
+        .iter()
+        .filter(|row| row.source_key == "child-1")
+        .collect();
+    assert_eq!(
+        parent_rows.len(),
+        2,
+        "the parent's original row and its appended row must both survive"
+    );
+    assert!(
+        child_rows.is_empty(),
+        "a source absent from this pass must not keep its rows forever"
+    );
+
+    let connection = store.lock();
+    assert_eq!(
+        query_source_resume(&connection, &turn_session_key(&key), "child-1").unwrap(),
+        None,
+        "a vanished source's stale resume snapshot must be dropped too"
     );
 }
